+++ packages/google/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ai-sdk/google
 
+## 4.0.0-beta.21
+
+### Patch Changes
+
+- Updated dependencies [2e17091]
+  - @ai-sdk/provider-utils@5.0.0-beta.9
+
+## 4.0.0-beta.20
+
+### Patch Changes
+
+- Updated dependencies [986c6fd]
+- Updated dependencies [493295c]
+  - @ai-sdk/provider-utils@5.0.0-beta.8
+
 ## 4.0.0-beta.19
 
 ### Patch Changes